requirements.txt
streamlit
polars
pandas
numpy
scikit-learn
langchain
langchain-openai
python-dotenv
tqdm
more-itertools
nest-asyncio
matplotlib 